LayoutTests/canvas/philip/tests/2d.line.miter.valid-expected.txt

< [index] >
2d.line.miter.valid
Setting miterLimit to valid values works
References: 2d.miterLimit.set 2d.miterLimit.get
Actual output:
Passed
